+++ src/backend/plugins/__init__.py
@@ -1,0 +1,1 @@
+"""Concrete implementation of plugins which can be used or not in the application."""