+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.9.0] - 2024-12-17
+
 ### Fixed
 
 - 🐛(backend) fix manage roles on domain admin view
@@ -220,7 +222,8 @@ and this project adheres to
 - ✨(domains) create and manage domains on admin + API
 - ✨(domains) mailbox creation + link to email provisioning API
 
-[unreleased]: https://github.com/numerique-gouv/people/compare/v1.8.0...main
+[unreleased]: https://github.com/numerique-gouv/people/compare/v1.9.0...main
+[1.9.0]: https://github.com/numerique-gouv/people/releases/v1.9.0
 [1.8.0]: https://github.com/numerique-gouv/people/releases/v1.8.0
 [1.7.1]: https://github.com/numerique-gouv/people/releases/v1.7.1
 [1.7.0]: https://github.com/numerique-gouv/people/releases/v1.7.0